extract user request mapper
