+++ intercom.go
@@ -12,6 +12,7 @@ type Client struct {
 	Contacts      ContactService
 	Conversations ConversationService
 	Events        EventService
+	Jobs          JobService
 	Messages      MessageService
 	Segments      SegmentService
 	Tags          TagService
@@ -23,6 +24,7 @@ type Client struct {
 	ContactRepository      ContactRepository
 	ConversationRepository ConversationRepository
 	EventRepository        EventRepository
+	JobRepository          JobRepository
 	MessageRepository      MessageRepository
 	SegmentRepository      SegmentRepository
 	TagRepository          TagRepository
@@ -101,6 +103,7 @@ func (c *Client) setup() {
 	c.ContactRepository = ContactAPI{httpClient: c.HTTPClient}
 	c.ConversationRepository = ConversationAPI{httpClient: c.HTTPClient}
 	c.EventRepository = EventAPI{httpClient: c.HTTPClient}
+	c.JobRepository = JobAPI{httpClient: c.HTTPClient}
 	c.MessageRepository = MessageAPI{httpClient: c.HTTPClient}
 	c.SegmentRepository = SegmentAPI{httpClient: c.HTTPClient}
 	c.TagRepository = TagAPI{httpClient: c.HTTPClient}
@@ -110,6 +113,7 @@ func (c *Client) setup() {
 	c.Contacts = ContactService{Repository: c.ContactRepository}
 	c.Conversations = ConversationService{Repository: c.ConversationRepository}
 	c.Events = EventService{Repository: c.EventRepository}
+	c.Jobs = JobService{Repository: c.JobRepository}
 	c.Messages = MessageService{Repository: c.MessageRepository}
 	c.Segments = SegmentService{Repository: c.SegmentRepository}
 	c.Tags = TagService{Repository: c.TagRepository}
